level-1/infra/pyproject.toml
[project]
name = "wio-terminal-infra"
version = "0.1.0"
description = "IoT AWS Infrastructure for WIO Terminal monitoring system"
requires-python = ">=3.9,<4"
dependencies = [
    "aws-cdk-lib==2.215.0",
    "constructs>=10.0.0,<11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "hypothesis>=6.0.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true